DONTMERGE: check_bump_tx: add a workaround for stripped witnesses in singlesig txs
get_raw_transaction_details() for singlesig returns the tx with the
witness data stripped. This means the tx cannot be re-broadcast by the
caller if lost, and the caller cannot verify the tx signatures.

As a temporary POC workaround until this is fixed, expect the full tx hex in
"previous_transaction/transaction" to allow bumping to correctly verify
the previous tx signatures.